feat: support multi users

+++ src/repositories/scheduler/StaffRepository.ts
@@ -26,13 +26,16 @@ class StaffRepository implements IStaffRepository {
 	 * @returns A promise that resolves to the found staff member or null if not found.
 	 * @throws {SqlReadError} If the query fails.
 	 */
-	async findByName(name: string): Promise<StaffDbModel | null> {
-		const query = 'SELECT * FROM Staffs WHERE name = ?';
+	async findByName(
+		name: string,
+		userId: string
+	): Promise<StaffDbModel | null> {
+		const query = 'SELECT * FROM Staffs WHERE name = ? AND userId = ?';
 
 		return (await this.queryService.getWithSqlErrorHandlingAsync(
 			this.databaseService.getDatabase(),
 			query,
-			[name],
+			[name, userId],
 			SqlReadError
 		)) as StaffDbModel | null;
 	}
@@ -43,15 +46,18 @@ class StaffRepository implements IStaffRepository {
 	 * @returns A promise that resolves to the created staff member.
 	 * @throws {SqlCreateError} If the query fails.
 	 */
-	async create(staff: StaffDbModel): Promise<StaffDbModel> {
-		const query =
-			'INSERT INTO Staffs (id, name, email, phoneNumber, image, color, createDate, modifyDate) VALUES (?, ?, ?, ?, ?, ?, ?, ?)';
+	async create(staff: StaffDbModel, userId: string): Promise<StaffDbModel> {
+		// const query =
+		// 	'INSERT INTO Staffs (id, name, email, phoneNumber, image, color, createDate, modifyDate) VALUES (?, ?, ?, ?, ?, ?, ?, ?)';
+
+		const query = `INSERT INTO Staffs (id, userId, name, email, phoneNumber, image, color, createDate, modifyDate, userId) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`;
 
 		await this.queryService.runWithSqlErrorHandlingAsync(
 			this.databaseService.getDatabase(),
 			query,
 			[
 				staff.id,
+				userId,
 				staff.name,
 				staff.email,
 				staff.phoneNumber,
@@ -71,13 +77,13 @@ class StaffRepository implements IStaffRepository {
 	 * @returns A promise that resolves to the found staff member or null if not found.
 	 * @throws {SqlReadError} If the query fails.
 	 */
-	async findById(id: string): Promise<StaffDbModel | null> {
-		const query = 'SELECT * FROM Staffs WHERE id = ?';
+	async findById(id: string, userId: string): Promise<StaffDbModel | null> {
+		const query = 'SELECT * FROM Staffs WHERE id = ? AND userId = ?';
 
 		return (await this.queryService.getWithSqlErrorHandlingAsync(
 			this.databaseService.getDatabase(),
 			query,
-			[id],
+			[id, userId],
 			SqlReadError
 		)) as StaffDbModel | null;
 	}
@@ -87,13 +93,13 @@ class StaffRepository implements IStaffRepository {
 	 * @returns A promise that resolves to an array of all staff members.
 	 * @throws {SqlReadError} If the query fails.
 	 */
-	async findAll(): Promise<StaffDbModel[]> {
-		const query = 'SELECT * FROM Staffs';
+	async findAll(userId: string): Promise<StaffDbModel[]> {
+		const query = 'SELECT * FROM Staffs WHERE userId = ?';
 
 		return (await this.queryService.allWithSqlErrorHandlingAsync(
 			this.databaseService.getDatabase(),
 			query,
-			[],
+			[userId],
 			SqlReadError
 		)) as StaffDbModel[];
 	}
@@ -104,9 +110,9 @@ class StaffRepository implements IStaffRepository {
 	 * @returns A promise that resolves to the updated staff member.
 	 * @throws {SqlUpdateError} If the query fails.
 	 */
-	async update(staff: StaffDbModel): Promise<StaffDbModel> {
+	async update(staff: StaffDbModel, userId: string): Promise<StaffDbModel> {
 		const query =
-			'UPDATE Staffs SET name = ?, email = ?, phoneNumber = ?, image = ?, color = ?, modifyDate = ? WHERE id = ?';
+			'UPDATE Staffs SET name = ?, email = ?, phoneNumber = ?, image = ?, color = ?, modifyDate = ? WHERE id = ? AND userId = ?';
 		await this.queryService.runWithSqlErrorHandlingAsync(
 			this.databaseService.getDatabase(),
 			query,
@@ -117,13 +123,14 @@ class StaffRepository implements IStaffRepository {
 				staff.image,
 				staff.color,
 				staff.modifyDate,
-				staff.id
+				staff.id,
+				userId
 			],
 			SqlUpdateError
 		);
 
 		// Return the updated staff
-		return (await this.findById(staff.id)) as StaffDbModel;
+		return (await this.findById(staff.id, userId)) as StaffDbModel;
 	}
 
 	/**
@@ -132,7 +139,7 @@ class StaffRepository implements IStaffRepository {
 	 * @returns A promise that resolves when the staff member is deleted.
 	 * @throws {SqlDeleteError} If the query fails.
 	 */
-	async deleteById(id: string): Promise<void> {
+	async deleteById(id: string, userId: string): Promise<void> {
 		// try {
 		// 	await this.findById(id);
 		// } catch (error) {
@@ -141,12 +148,12 @@ class StaffRepository implements IStaffRepository {
 		// 	}
 		// }
 
-		const query = 'DELETE FROM Staffs WHERE id = ?';
+		const query = 'DELETE FROM Staffs WHERE id = ? AND userId = ?';
 
 		await this.queryService.runWithSqlErrorHandlingAsync(
 			this.databaseService.getDatabase(),
 			query,
-			[id],
+			[id, userId],
 			SqlDeleteError
 		);
 	}
@@ -157,13 +164,13 @@ class StaffRepository implements IStaffRepository {
 	 * @returns A promise that resolves when the staff member is deleted.
 	 * @throws {SqlDeleteError} If the query fails.
 	 */
-	async deleteByName(name: string): Promise<void> {
-		const query = 'DELETE FROM Staffs WHERE name = ?';
+	async deleteByName(name: string, userId: string): Promise<void> {
+		const query = 'DELETE FROM Staffs WHERE name = ? AND userId = ?';
 
 		await this.queryService.runWithSqlErrorHandlingAsync(
 			this.databaseService.getDatabase(),
 			query,
-			[name],
+			[name, userId],
 			SqlDeleteError
 		);
 	}
